Fixed misc code review issues.
