+++ Corale.Colore/Core/Mousepad.Obsoletes.cs
@@ -1,5 +1,5 @@
 ﻿// ---------------------------------------------------------------------------------------
-// <copyright file="SimpleHeadset.cs" company="Corale">
+// <copyright file="SimpleMousepad.cs" company="Corale">
 //     Copyright © 2015 by Adam Hellberg and Brandon Scott.
 //
 //     Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -31,49 +31,62 @@
 namespace Corale.Colore.Core
 {
     using System;
-    using Corale.Colore.Razer.Headset.Effects;
+    using Corale.Colore.Razer.Mousepad.Effects;
 
     /// <summary>
-    /// Class for interacting with Chroma Headsets.
+    /// Class for interacting with a Chroma mouse pad.
     /// </summary>
-    public sealed partial class Headset : Device, IHeadset
+    public sealed partial class Mousepad
     {
         /// <summary>
-        /// Sets an effect on the headset that doesn't
-        /// take any parameters, currently only valid
-        /// for the <see cref="Effect.SpectrumCycling" /> effect.
+        /// Sets an effect without any parameters.
+        /// Currently, this only works for the <see cref="Effect.None" /> effect.
         /// </summary>
-        /// <param name="effect">The type of effect to set.</param>
+        /// <param name="effect">Effect options.</param>
         [Obsolete("Set is deprecated, please use SetEffect.", false)]
         public void Set(Effect effect)
         {
-            SetEffect(effect);
+            SetGuid(NativeWrapper.CreateMousepadEffect(effect));
         }
 
         /// <summary>
-        /// Sets a new static effect on the headset.
+        /// Sets a breathing effect on the mouse pad.
         /// </summary>
-        /// <param name="effect">
-        /// An instance of the <see cref="Static" /> struct
-        /// describing the effect.
-        /// </param>
-        [Obsolete("Set is deprecated, please use SetStatic.", false)]
-        public void Set(Static effect)
-        {
-            SetStatic(effect);
-        }
-
-        /// <summary>
-        /// Sets a new breathing effect on the headset.
-        /// </summary>
-        /// <param name="effect">
-        /// An instance of the <see cref="Breathing" /> struct
-        /// describing the effect.
-        /// </param>
+        /// <param name="effect">An instance of the <see cref="Breathing" /> struct.</param>
         [Obsolete("Set is deprecated, please use SetBreathing.", false)]
         public void Set(Breathing effect)
         {
-            SetBreathing(effect);
+            SetGuid(NativeWrapper.CreateMousepadEffect(effect));
+        }
+
+        /// <summary>
+        /// Sets a static color effect on the mouse pad.
+        /// </summary>
+        /// <param name="effect">An instance of the <see cref="Static" /> struct.</param>
+        [Obsolete("Set is deprecated, please use SetStatic.", false)]
+        public void Set(Static effect)
+        {
+            SetGuid(NativeWrapper.CreateMousepadEffect(effect));
+        }
+
+        /// <summary>
+        /// Sets a wave effect on the mouse pad.
+        /// </summary>
+        /// <param name="effect">An instance of the <see cref="Wave" /> struct.</param>
+        [Obsolete("Set is deprecated, please use SetWave.", false)]
+        public void Set(Wave effect)
+        {
+            SetGuid(NativeWrapper.CreateMousepadEffect(effect));
+        }
+
+        /// <summary>
+        /// Sets a custom effect on the mouse pad.
+        /// </summary>
+        /// <param name="effect">An instance of the <see cref="Custom" /> struct.</param>
+        [Obsolete("Set is deprecated, please use SetCustom.", false)]
+        public void Set(Custom effect)
+        {
+            SetGuid(NativeWrapper.CreateMousepadEffect(effect));
         }
     }
 }
